fix: PlusButton
1. PlusButton을 AppLayout에서 제외하고 사용할 페이지에서만 추가
2. 스타일을 PlusButton 폴더 아래에 생성하고 AppLayout에서 삭제
#14
