+++ src/components/atomic/atoms/PlusButton/index.tsx
@@ -4,17 +4,21 @@ import { Link } from 'react-router-dom';
 import Icon from '@/components/atomic/atoms/Icon';
 import Button from '@/components/atomic/atoms/Button';
 
+import { PositionWrapper } from './style';
+
 const PlusButton = () => {
   return (
     <Link to="/group/create">
-      <Button
-        backgroundColor="#FFFFFF"
-        borderRadius="50%"
-        width="40px"
-        height="40px"
-        boxShadow="0px 1px 4px rgba(62, 62, 62, 0.15)">
-        <Icon icon="Plus" />
-      </Button>
+      <PositionWrapper>
+        <Button
+          backgroundColor="#FFFFFF"
+          borderRadius="50%"
+          width="40px"
+          height="40px"
+          boxShadow="0px 1px 4px rgba(62, 62, 62, 0.15)">
+          <Icon icon="Plus" />
+        </Button>
+      </PositionWrapper>
     </Link>
   );
 };
