+++ src/components/common/AppLayout/index.tsx
@@ -3,11 +3,9 @@ import React from 'react';
 import Header from '@/components/common/Header';
 import Footer from '@/components/common/Footer';
 
-import PlusButton from '@/components/atomic/atoms/PlusButton';
-
 import NavBar from '@/components/atomic/molecules/NavBar';
 
-import { Wrapper, ContentsWrapper, Contents, PositionWrapper } from './style';
+import { Wrapper, ContentsWrapper, Contents } from './style';
 
 interface IAppLayoutComponentProps {
   children: React.ReactNode;
@@ -21,9 +19,6 @@ const AppLayout = ({ children }: IAppLayoutComponentProps) => {
       <ContentsWrapper>
         <Contents>{children}</Contents>
       </ContentsWrapper>
-      <PositionWrapper>
-        <PlusButton />
-      </PositionWrapper>
       <Footer />
     </Wrapper>
   );
