+++ src/components/common/AppLayout/style.ts
@@ -1,8 +1,5 @@
 import styled from 'styled-components';
 
-const { width, height } = screen;
-const halfMargin = Math.floor((width - 790) / 2);
-
 export const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -23,16 +20,5 @@ export const Contents = styled.div`
 
   ${({ theme }) => theme.desktop} {
     width: 790px;
-  }
-`;
-
-export const PositionWrapper = styled.div`
-  position: fixed;
-  top: ${height - 50}px;
-  left: ${width - 50}px;
-
-  ${({ theme }) => theme.desktop} {
-    top: ${height - 200}px;
-    left: ${width - halfMargin}px;
   }
 `;
